+++ tiledb/sm/filter/filter.cc
@@ -31,6 +31,7 @@
  */
 
 #include "tiledb/sm/filter/filter.h"
+#include "tiledb/sm/filter/bit_width_reduction_filter.h"
 #include "tiledb/sm/filter/compression_filter.h"
 #include "tiledb/sm/filter/noop_filter.h"
 #include "tiledb/sm/misc/logger.h"
@@ -68,6 +69,8 @@ Filter* Filter::create(FilterType type) {
     case FilterType::FILTER_BZIP2:
     case FilterType::FILTER_DOUBLE_DELTA:
       return new (std::nothrow) CompressionFilter(type, -1);
+    case FilterType::FILTER_BIT_WIDTH_REDUCTION:
+      return new (std::nothrow) BitWidthReductionFilter();
     default:
       assert(false);
       return nullptr;